requirements.txt
asgiref==3.5.0
Django==4.0.1
djangorestframework==3.13.1
isort==5.10.1
pytz==2021.3
sqlparse==0.4.2
